+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -1,4 +1,5 @@
 import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
+import { BusinessReportService } from '@/business-report/business-report.service';
 import { BusinessRepository } from '@/business/business.repository';
 import { BusinessService } from '@/business/business.service';
 import { ajv } from '@/common/ajv/ajv.validator';
@@ -20,7 +21,10 @@ import {
 } from '@/prisma/prisma.util';
 import { ProjectScopeService } from '@/project/project-scope.service';
 import { FileService } from '@/providers/file/file.service';
+import { RiskRuleService, TFindAllRulesOptions } from '@/rule-engine/risk-rule.service';
+import { RuleEngineService } from '@/rule-engine/rule-engine.service';
 import { SalesforceService } from '@/salesforce/salesforce.service';
+import { SentryService } from '@/sentry/sentry.service';
 import type {
   InputJsonValue,
   IObjectWithId,
@@ -82,7 +86,9 @@ import {
   WorkflowRuntimeData,
   WorkflowRuntimeDataStatus,
 } from '@prisma/client';
+import { Static } from '@sinclair/typebox';
 import { plainToClass } from 'class-transformer';
+import dayjs from 'dayjs';
 import { isEqual, merge } from 'lodash';
 import mime from 'mime';
 import { WorkflowDefinitionCreateDto } from './dtos/workflow-definition-create';
@@ -97,15 +103,9 @@ import {
   WorkflowRuntimeListQueryResult,
 } from './types';
 import { addPropertiesSchemaToDocument } from './utils/add-properties-schema-to-document';
+import { entitiesUpdate } from './utils/entities-update';
 import { WorkflowEventEmitterService } from './workflow-event-emitter.service';
 import { WorkflowRuntimeDataRepository } from './workflow-runtime-data.repository';
-import { Static } from '@sinclair/typebox';
-import dayjs from 'dayjs';
-import { entitiesUpdate } from './utils/entities-update';
-import { BusinessReportService } from '@/business-report/business-report.service';
-import { RuleEngineService } from '@/rule-engine/rule-engine.service';
-import { RiskRuleService, TFindAllRulesOptions } from '@/rule-engine/risk-rule.service';
-import { SentryService } from '@/sentry/sentry.service';
 
 type TEntityId = string;
 
@@ -1950,7 +1950,6 @@ export class WorkflowService {
           machineContext: workflowRuntimeData.context,
           state: workflowRuntimeData.state,
         },
-        // @ts-expect-error - error from Prisma types fix
         extensions: workflowDefinition.extensions,
         invokeRiskRulesAction: async (
           context: object,